+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Parable GetSet
 
+## 0.1.1
+
+_Changes_
+- Add `string` type-hinting to `$key` for `get()` and `getAndRemove()`, in line with the rest.
+- Add `$default = null` parameter to `get()`, so that the scenarios of 'not set' and 'set but `null`' can be distinguished between.
+
 ## 0.1.0
 
 _Changes_